connection.py
import os
from dotenv import load_dotenv
from sqlmodel import SQLModel, Session, create_engine

load_dotenv()
username = os.getenv("DB_USERNAME", "postgres")
password = os.getenv("DB_PASSWORD", "")
host = os.getenv("DB_HOST", "localhost")
database = os.getenv("DB_NAME", "web_team_finder")

db_url = f'postgresql://{username}:{password}@{host}/{database}'
engine = create_engine(db_url, echo=True)


def init_db():
    SQLModel.metadata.create_all(engine)


def get_session():
    with Session(engine) as session:
        yield session
