pyproject.toml
[project]
name = "itmo-ict-webdevelopment-tools-2024-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "bs4>=0.0.2",
    "fastapi[all]>=0.115.12",
    "jwt>=1.3.1",
    "lxml>=5.4.0",
    "matplotlib>=3.10.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pymdown-extensions>=10.15",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlmodel>=0.0.24",
]
